Add spisoft_arch.h
